Refactored to avoid split packages
