+++ examples/rest-assured-itest-java/src/test/java/io/restassured/itest/java/CustomObjectMappingITest.java
@@ -26,7 +26,8 @@ import io.restassured.itest.java.support.WithJetty;
 import io.restassured.mapper.ObjectMapper;
 import io.restassured.mapper.ObjectMapperDeserializationContext;
 import io.restassured.mapper.ObjectMapperSerializationContext;
-import io.restassured.mapper.factory.GsonObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.GsonObjectMapperFactory;
+
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Before;
 import org.junit.Test;
