+++ json-path/src/main/java/io/restassured/path/json/config/JsonPathConfig.java
@@ -16,7 +16,13 @@
 
 package io.restassured.path.json.config;
 
-import io.restassured.mapper.factory.*;
+import io.restassured.common.mapper.factory.*;
+import io.restassured.path.json.mapper.factory.DefaultGsonObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.DefaultJackson1ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.DefaultJackson2ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.GsonObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.Jackson1ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.Jackson2ObjectMapperFactory;
 import io.restassured.path.json.mapping.JsonPathObjectDeserializer;
 import org.apache.commons.lang3.StringUtils;
