+++ rest-assured-common/src/main/java/io/restassured/common/mapper/TypeRef.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.restassured.mapper;
+package io.restassured.common.mapper;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
