+++ rest-assured-common/src/main/java/io/restassured/common/mapper/resolver/ObjectMapperResolver.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.restassured.mapper.resolver;
+package io.restassured.common.mapper.resolver;
 
 import static io.restassured.internal.classpath.ClassPathResolver.existInCP;
