+++ rest-assured/src/main/java/io/restassured/config/ObjectMapperConfig.java
@@ -18,7 +18,14 @@ package io.restassured.config;
 
 import io.restassured.mapper.ObjectMapper;
 import io.restassured.mapper.ObjectMapperType;
-import io.restassured.mapper.factory.*;
+import io.restassured.path.json.mapper.factory.DefaultGsonObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.DefaultJackson1ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.DefaultJackson2ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.GsonObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.Jackson1ObjectMapperFactory;
+import io.restassured.path.json.mapper.factory.Jackson2ObjectMapperFactory;
+import io.restassured.path.xml.mapper.factory.*;
+
 import org.apache.commons.lang3.Validate;
 
 /**
