+++ rest-assured/src/main/java/io/restassured/internal/RestAssuredResponseOptionsImpl.java
@@ -16,6 +16,7 @@
 
 package io.restassured.internal;
 
+import io.restassured.common.mapper.TypeRef;
 import io.restassured.config.DecoderConfig;
 import io.restassured.config.RestAssuredConfig;
 import io.restassured.http.Cookie;
@@ -24,7 +25,6 @@ import io.restassured.http.Headers;
 import io.restassured.internal.log.LogRepository;
 import io.restassured.mapper.ObjectMapper;
 import io.restassured.mapper.ObjectMapperType;
-import io.restassured.mapper.TypeRef;
 import io.restassured.path.json.JsonPath;
 import io.restassured.path.json.config.JsonPathConfig;
 import io.restassured.path.xml.XmlPath;
