+++ rest-assured/src/main/java/io/restassured/response/ResponseBodyExtractionOptions.java
@@ -16,9 +16,9 @@
 
 package io.restassured.response;
 
+import io.restassured.common.mapper.TypeRef;
 import io.restassured.mapper.ObjectMapper;
 import io.restassured.mapper.ObjectMapperType;
-import io.restassured.mapper.TypeRef;
 import io.restassured.path.json.JsonPath;
 import io.restassured.path.json.config.JsonPathConfig;
 import io.restassured.path.xml.XmlPath;
