+++ xml-path/pom.xml
@@ -38,7 +38,7 @@
                 <extensions>true</extensions>
                 <configuration>
                     <instructions>
-                        <Export-Package>io.restassured.mapper.*, io.restassured.path.xml.*</Export-Package>
+                        <Export-Package>io.restassured.path.xml.*</Export-Package>
                         <Import-Package>
 							groovy.*;version="${groovy.range}",
                             org.codehaus.groovy.*;version="${groovy.range}",
